events: split EventManager to sub-package
