+++ events/scheduler.go
@@ -1,0 +1,10 @@
+package events
+
+import (
+	"context"
+)
+
+type Scheduler interface {
+	AddWork(ctx context.Context, repo string, val *XRPCStreamEvent) error
+	Shutdown()
+}
